Rename notifications to notification_service
